ci: VK Teams instead of Telegram for notifications
This patch introduces the new `report-job-status` action that is used
for sending notifications to the VK Teams messenger on the job failure.
It replaces the old `send-telegram-notify` action sending similar text
notifications to the Telegram messenger.

The new action selects the corresponding chat for the notification on
failure and uses the `tarantool/actions/report-job-status` action [1]
for sending it to:

* the team chat (Tarantool CI/CD reports) - if the job started on any
  event in the master or release branch (1.10, 2.10, etc);

* a personal chat (created by the committer) – if the job started on
  creating/updating a pull request or any event in other branches.

[1] https://github.com/tarantool/actions/tree/master/report-job-status

Resolves tarantool/tarantool-qa#258

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci