crash: expand log message about crashdump
Crashdump is written to debug log before it is sent
to feedback daemon. The problem is that the user may
see this message and think that crashdump is logged
for him, while it is logged only for debug of feedback
daemon. This patch adds an explanation to log message
with coredump.

Closes tarantool/tarantool-ee#199

NO_DOC=change log message
NO_TEST=change log message