index: add index_read_view get method
We will use it in the read view Lua API. For memtx tree and hash
indexes, we add a stub function with a possibility of override if
ENABLE_READ_VIEW is set. For the sequence data space, we raise an
error if this function is called, because we aren't planning to
implement it in the EE repository.

Needed for https://github.com/tarantool/tarantool-ee/issues/197

NO_DOC=internal
NO_TEST=ee
NO_CHANGELOG=internal