sql: introduce new parsing rule
This commit introduces a new parse rule for compiling an unresolved
single expression. This simplifies the current implementation of
sql_expr_compile() and is useful for generating SQL expressions that can
be used in the core check constraint. This rule is for internal use
only.

Part of #6986

NO_DOC=will be added later
NO_TEST=refactoring
NO_CHANGELOG=will be added later