memtx: allow to override read view iterator implementation
For memtx tree and hash indexes: if the ENABLE_READ_VIEW macro is
defined, include a source file with the read view iterator
implementation instead of defining it directly. We don't need to support
any iterator types except ALL in the CE repository, because user read
views will be available only in EE. All other iterator types will be
defined in the EE repository. We include source files (not headers),
because to implement a read view iterator, we need access to index
internals, which are defined in memtx_hash.cc and memtx_tree.cc.

The included implementation source file is supposed to define
the {tree,hash}_read_view_iterator_start function, which positions the
iterator to the given key. It's called directly by create_iteartor index
read view method.

For the memtx tree index, the iterator implementation needs to know the
iteration key (for handling EQ/REQ requests) so we also add the key to
the iterator struct in this commit.

Needed for https://github.com/tarantool/tarantool-ee/issues/197

NO_DOC=internal
NO_TEST=ee
NO_CHANGELOG=internal