read_view: assert that read view is used only in one thread
Although currenlty feasible, using a read view from multiple threads
simultaneously is dangerous on its own because of inevitably arising
object life time issues. With introduction of space upgrade handling,
it'll become hardly possible, because we'll have to attach Lua function
to a read view, which can only be used from one thread. Let's add some
assertions to guarantee that a read view is used only from one thread.
To do that, we need to introduce the concept of an active read view:
from now on, a read view must be activated before it can be used and
deactivated before it can be closed; both operations must be done from
the same thread - the thread that accesses the read view.

https://github.com/tarantool/tarantool-ee/issues/163

NO_DOC=debug
NO_TEST=debug
NO_CHANGELOG=debug