error: add HandleIsClosed error
It is suitable for situations, when a handle is closed during a long
(yielding) operation: say, by another fiber.

Unlike ChannelIsClosed it is not specific for the fiber channel and
allows to clarify what is going on in the error message.

Usage of ChannelIsClosed for anything expect fiber channels (or at least
anything that does not feel like a channel) would be confusing.

Part of #7653

NO_DOC=The documentation does not list all errors currently and there
       is no much need to create such list: the errors are primarily to
       be handled by a user and not to be created in user's code. It
       looks okay to just mention the error type a description of an API
       that returns/raises the error. I'll do it in the next commit.
NO_CHANGELOG=see NO_DOC