replication: introduce vclock sync requests
Currently replication acks deliver vclock the replica has reached to the
master. But there is no guarantee on how fresh that vclock is: by the
time the ack arrives, real replica vclock might be much greater.

In some applications relay can't rely on arbitrarily old ack vclock. It
needs an upper bound on vclock which could be reached by the replica at
some specific point in time.

In order to achieve this, introduce vclock syncs. A vclock sync is a
monotonically growing heartbeat identifier, unique for a live
master-replica connection. This identifier may now be appended to
heartbeats coming from the master. Once receiving a heartbeat with some
non-zero vclock sync value, replica starts responding with the same
vclock sync value in all its acks.

Here's how it can be used: for example, in multi-master configuration,
one master wants to wait until it receives all the data, that was present
on a remote master at some point in time. In order to do so, the first
master issues a vclock sync request. Once it receives an ack with the
same vclock sync it has sent, it knows for sure that the remote master's
vclock was not greater than the one received in the ack.

This may be used to implement linearizable reads.

Prerequisite #6707

NO_TEST=tested in next commit
NO_CHANGELOG=internal change

@TarantoolBot document
Title: Binary protocol: new key - IPROTO_VCLOCK_SYNC
Binary protocol receives a new key:
IPROTO_VCLOCK_SYNC = 0x5a
This key holds a MP_UINT value and is used only by replication
heartbeats. A master sends the monotonically growing vclock sync together
with some of its heartbeats, and replica replies with the greatest vclock
sync it has seen yet in its ACKs.